+++ R/config.R
@@ -149,12 +149,7 @@ py_discover_config <- function(required_module = NULL, use_environment = NULL) {
   main_process_info <- main_process_python_info()
   if (!is.null(main_process_info)) {
     python_version <- normalize_python_path(main_process_info$python)$path
-    config <- python_config(
-      python_version,
-      required_module,
-      python_version,
-      forced = "the current process"
-    )
+    config <- python_config(python_version, required_module, forced = "the current process")
     return(config)
   }
 
@@ -164,38 +159,53 @@ py_discover_config <- function(required_module = NULL, use_environment = NULL) {
   py_session_initialized <- py_session_initialized_binary()
   if (!is.null(py_session_initialized)) {
     python_version <- normalize_python_path(py_session_initialized)$path
-    config <- python_config(python_version, required_module, python_version, forced = "PYTHON_SESSION_INITIALIZED")
+    config <- python_config(python_version, required_module, forced = "PYTHON_SESSION_INITIALIZED")
     return(config)
   }
 
   # if RETICULATE_PYTHON is specified then use it without scanning further
   reticulate_env <- Sys.getenv("RETICULATE_PYTHON", unset = NA)
   if (!is.na(reticulate_env)) {
+    
     python_version <- normalize_python_path(reticulate_env)
     if (!python_version$exists)
       stop("Python specified in RETICULATE_PYTHON (", reticulate_env, ") does not exist")
+    
     python_version <- python_version$path
-    config <- python_config(python_version, required_module, python_version, forced = "RETICULATE_PYTHON")
+    config <- python_config(python_version, required_module, forced = "RETICULATE_PYTHON")
     return(config)
+    
   }
 
   # if RETICULATE_PYTHON_ENV is specified then use that
   reticulate_python_env <- Sys.getenv("RETICULATE_PYTHON_ENV", unset = NA)
   if (!is.na(reticulate_python_env)) {
+    
     python <- python_binary_path(reticulate_python_env)
     python_version <- normalize_python_path(python)
     if (!python_version$exists)
       stop("Python specified in RETICULATE_PYTHON_ENV (", reticulate_python_env, ") does not exist")
+    
     path <- python_version$path
-    config <- python_config(path, required_module, path, forced = "RETICULATE_PYTHON_ENV")
+    config <- python_config(path, required_module, forced = "RETICULATE_PYTHON_ENV")
+    return(config)
+    
+  }
+  
+  # if we're working within a project that contains a Pipfile, then
+  # use the copy of Python associated with that pipenv
+  pipfile <- pipenv_pipfile_path()
+  if (file.exists(pipfile)) {
+    python <- pipenv_python()
+    config <- python_config(python, required_module, forced = "Pipfile")
     return(config)
   }
-
+  
   # next look for a required python version (e.g. use_python("/usr/bin/python", required = TRUE))
   required_version <- .globals$required_python_version
   if (!is.null(required_version)) {
     python_version <- normalize_python_path(required_version)$path
-    config <- python_config(python_version, required_module, python_version, forced = "use_python function")
+    config <- python_config(python_version, required_module, forced = "use_python function")
     return(config)
   }
   
@@ -511,7 +521,10 @@ python_munge_path <- function(python) {
 
 }
 
-python_config <- function(python, required_module, python_versions, forced = NULL) {
+python_config <- function(python,
+                          required_module,
+                          python_versions = python,
+                          forced = NULL) {
 
   # normalize and remove duplicates
   python <- canonical_path(python)